Move code from `pkg/maintainer/wallet` to `pkg/tbtcpg` package
This move is meant to reflect the fact that wallet maintainer is no longer
a thing and, we want to make proposal generator a standalone module that will
be used by `pkg/tbtc`. By the way, we are also cleaning up proposal-related
commands from the maintainer-cli.
