--- pkg/maintainer/wallet/wallet.go
@@ -1,7 +1,0 @@
-package wallet
-
-import (
-	"github.com/ipfs/go-log/v2"
-)
-
-var logger = log.Logger("keep-maintainer-wallet")
