+++ pkg/tbtcpg/bitcoin_chain_test.go
@@ -1,4 +1,4 @@
-package wallet
+package tbtcpg
 
 import (
 	"fmt"
