+++ pkg/tbtcpg/chain_test.go
@@ -1,4 +1,4 @@
-package wallet
+package tbtcpg
 
 import (
 	"bytes"
