+++ pkg/tbtcpg/internal/test/testdata/propose_sweep_scenario_0.json
