+++ pkg/tbtcpg/internal/test/testdata/propose_sweep_scenario_2.json
