+++ pkg/tbtcpg/redemptions.go
@@ -1,4 +1,4 @@
-package wallet
+package tbtcpg
 
 import (
 	"encoding/hex"
