+++ pkg/tbtcpg/tbtcpg.go
@@ -1,0 +1,7 @@
+package tbtcpg
+
+import (
+	"github.com/ipfs/go-log/v2"
+)
+
+var logger = log.Logger("keep-tbtcpg")
